backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
firebase-admin==6.4.0
pinecone-client==3.0.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
langchain==0.1.9
langchain-community==0.0.19
langchain-core==0.1.27
deepseek-ai==0.1.0
openai==1.12.0
numpy==1.26.3